# server/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/users/{id}:
    get:
      tags:
      - user-rest-controller
      summary: Get user by ID
      operationId: getUser
      parameters:
      - name: id
        in: path
        description: ID of the user to retrieve
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        description: Pageable object for pagination
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: Access Denied
        "200":
          description: Return the user with the ID given
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "404":
          description: The requested resource could not be found
    put:
      tags:
      - user-rest-controller
      summary: Update user by ID
      operationId: putUser
      parameters:
      - name: id
        in: path
        description: ID of the user to update
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: Access Denied
        "200":
          description: User modified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserNameResponse'
        "404":
          description: The requested resource could not be found
    delete:
      tags:
      - user-rest-controller
      summary: Delete user by ID
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        description: ID of the user to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserNameResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: Access Denied
        "404":
          description: The requested resource could not be found
...
  /api/auth/refresh:
    post:
      tags:
      - login-controller
      summary: Refresh token
      operationId: refreshToken
      parameters:
      - name: refreshToken
        in: cookie
        description: Refresh token to refresh authentication
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Token refreshed successful
          content:
            application/json: {}
...
  /api/users/:
    get:
      tags:
      - user-rest-controller
      summary: Get all users
      operationId: getAllUser
      parameters:
      - name: page
        in: query
        description: Pageable object for pagination
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: Return the list of user with their roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserIdAndNameAndRolesResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: Access Denied
...
components:
  schemas:
    UserNameResponse:
      required:
      - Name
      type: object
      properties:
        Name:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    AerogeneratorResponse:
      required:
      - aerogeneratorId
      - aerogeneratorLatitude
      - aerogeneratorLongitude
      - aerogeneratorPower
      - bladeLength
      - height
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        aerogeneratorId:
          type: string
        aerogeneratorLatitude:
          type: number
          format: double
        aerogeneratorLongitude:
          type: number
          format: double
        bladeLength:
          type: integer
          format: int32
        height:
          type: integer
          format: int32
        aerogeneratorPower:
          type: integer
          format: int32
    EoloparkResponse:
      required:
      - aerogenerators
      - area
      - city
      - id
      - latitude
      - longitude
      - name
      - substation
      - terrainType
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        city:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        area:
          type: integer
          format: int32
        terrainType:
          type: string
        substation:
          $ref: '#/components/schemas/SubstationResponse'
        aerogenerators:
          type: array
          items:
            $ref: '#/components/schemas/AerogeneratorResponse'
    SubstationResponse:
      required:
      - model
      - power
      - voltage
      type: object
      properties:
        model:
          type: string
        power:
          type: number
          format: double
        voltage:
          type: number
          format: double
    UserResponse:
      required:
      - id
      - name
      - parkMax
      - roles
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        parkMax:
          type: integer
          format: int32
        roles:
          type: array
          items:
            type: string
        eoloparks:
          type: array
          items:
            $ref: '#/components/schemas/EoloparkResponse'
    UserIdAndNameAndRolesResponse:
      required:
      - id
      - name
      - roles
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        roles:
          type: array
          items:
            type: string
...